rename TestTests to Tests
