+++ Tests/PaymentArea/PaymentTests.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace moyasar.PaymentArea.Tests
+namespace Moyasar.Tests.PaymentArea
 {
     [TestClass()]
     public class PaymentTests
