+++ Tests/Tests.csproj
@@ -6,8 +6,8 @@
     <ProjectGuid>{692C1BF7-23D8-438D-8FB3-280962833B48}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>TestTests</RootNamespace>
-    <AssemblyName>TestTests</AssemblyName>
+    <RootNamespace>Moyasar.Tests</RootNamespace>
+    <AssemblyName>MoyasarTests</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{3AC096D0-A1C2-E12C-1390-A8335801FDAB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
@@ -51,7 +51,7 @@
     </Otherwise>
   </Choose>
   <ItemGroup>
-    <Compile Include="Class1Tests.cs" />
+    <Compile Include="Tests.cs" />
     <Compile Include="PaymentArea\PaymentTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
@@ -59,10 +59,6 @@
     <ProjectReference Include="..\moyasar\moyasar.csproj">
       <Project>{9EC79375-DF6A-4E74-A90C-731751711E0D}</Project>
       <Name>moyasar</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\Test\Test.csproj">
-      <Project>{09445DCE-DB47-4086-9779-BEE24C754339}</Project>
-      <Name>Test</Name>
     </ProjectReference>
   </ItemGroup>
   <Choose>
